Extracted the pull mode part into a separate handler
